worker: refactor: split build metrics and drv translation out of build.rs and unblock the async paths
