+++ backend/gradient-worker/src/executor/build.rs
@@ -21,265 +21,32 @@
 //! together and reports the result to the server.
 
 use anyhow::{Context, Result};
-use bytes::Bytes;
 use futures::StreamExt as _;
-use gradient_db::{DrvOutputSpec, parse_drv};
-use gradient_exec::path_utils::{nix_store_path, strip_nix_store_prefix, strip_store_prefix};
-use gradient_proto::messages::{BuildMetrics, BuildOutput, BuildProduct, BuildTask};
+use gradient_db::parse_drv;
+use gradient_exec::path_utils::{nix_store_path, strip_store_prefix};
+use gradient_proto::messages::{BuildOutput, BuildProduct, BuildTask};
 use gradient_sources::get_hash_from_path;
 use gradient_util::hydra::parse_hydra_product_line;
-use harmonia_protocol::daemon_wire::types2::{BuildMode, BuildResultInner, Microseconds};
+use harmonia_protocol::daemon_wire::types2::{BuildMode, BuildResultInner};
 use harmonia_protocol::log::{Field, LogMessage, ResultType, Verbosity};
 use harmonia_protocol::types::ClientOptions;
-use harmonia_store_content_address::{ContentAddress, ContentAddressMethod};
-use harmonia_store_derivation::derivation::{BasicDerivation, DerivationOutput, DerivationT};
+use harmonia_store_derivation::derivation::BasicDerivation;
 use harmonia_store_path::StorePath;
 use harmonia_store_remote::DaemonStore as _;
-use harmonia_utils_hash::{Algorithm, Hash};
 use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
 use std::pin::{Pin, pin};
 use tokio::sync::watch;
 use tracing::{debug, info, warn};
 
-use crate::metrics::cgroup::{BuildMetricsRaw, read_build_cgroup};
 use crate::nix::store::LocalNixStore;
 use crate::proto::job::JobUpdater;
 
+use super::build_metrics::{
+    CgroupSampler, NetworkPeakSampler, assemble_build_metrics, daemon_cpu_usec,
+};
+use super::derivation::get_basic_derivation;
 pub use super::failure::BuildError;
 use super::failure::classify_build_error;
-
-const BYTES_PER_MB: u64 = 1_048_576;
-/// Cadence for sampling a live build cgroup's `memory.peak` / `io.stat`.
-const CGROUP_SAMPLE_MS: u64 = 200;
-
-// ── Build metrics ───────────────────────────────────────────────────────────
-
-/// Convert a raw cgroup snapshot into the wire `BuildMetrics`.
-///
-/// `build_time_ms` is always reported. Cgroup-derived fields are `None` when
-/// `raw` is absent. `avg_cpu_pct` is `None` when it cannot be computed
-/// (zero build time or zero CPU count) to avoid divide-by-zero.
-fn raw_to_build_metrics(
-    raw: Option<BuildMetricsRaw>,
-    build_time_ms: u64,
-    cpu_count: u32,
-    peak_network_mbps: Option<f32>,
-) -> BuildMetrics {
-    let Some(raw) = raw else {
-        return BuildMetrics {
-            build_time_ms: Some(build_time_ms),
-            peak_network_mbps,
-            ..Default::default()
-        };
-    };
-
-    let cpu_time_ms = raw.cpu_usage_usec.map(|u| u / 1000);
-    let avg_cpu_pct = match cpu_time_ms {
-        Some(cpu_ms) if build_time_ms > 0 && cpu_count > 0 => {
-            Some(cpu_ms as f32 / (build_time_ms as f32 * cpu_count as f32) * 100.0)
-        }
-        _ => None,
-    };
-
-    BuildMetrics {
-        peak_ram_mb: raw.peak_ram_bytes.map(|b| b / BYTES_PER_MB),
-        cpu_time_ms,
-        avg_cpu_pct,
-        disk_read_bytes: Some(raw.disk_read_bytes),
-        disk_write_bytes: Some(raw.disk_write_bytes),
-        oom_killed: raw.oom_killed,
-        build_time_ms: Some(build_time_ms),
-        peak_network_mbps,
-    }
-}
-
-/// Locate a running build's cgroup via nix's `<state-dir>/cgroups/` map.
-///
-/// Nix can't be asked which build user / cgroup it assigned, but on each build
-/// it writes the build's absolute cgroup path to `<state-dir>/cgroups/<uid>`.
-/// Those files persist after the build (pointing at a since-destroyed cgroup),
-/// so we only consider entries modified at/after `since` (the build's start)
-/// and return the newest. Returns `None` when nothing is newer than `since`
-/// (idle, daemon not using cgroups, or concurrent starts we won't guess at).
-fn newest_build_cgroup(state_dir: &Path, since: std::time::SystemTime) -> Option<PathBuf> {
-    let mut newest: Option<(std::time::SystemTime, PathBuf)> = None;
-    for entry in std::fs::read_dir(state_dir).ok()?.flatten() {
-        let Ok(mtime) = entry.metadata().and_then(|m| m.modified()) else {
-            continue;
-        };
-        if mtime < since {
-            continue;
-        }
-        if newest.as_ref().is_none_or(|(t, _)| mtime > *t)
-            && let Ok(contents) = std::fs::read_to_string(entry.path())
-        {
-            let path = PathBuf::from(contents.trim());
-            if !path.as_os_str().is_empty() {
-                newest = Some((mtime, path));
-            }
-        }
-    }
-    newest.map(|(_, p)| p)
-}
-
-/// Total CPU microseconds from the daemon's `BuildResult` (cgroup-derived,
-/// captured by nix before it tears the cgroup down). `None` only when the
-/// daemon reported neither user nor system time.
-fn daemon_cpu_usec(user: Option<Microseconds>, system: Option<Microseconds>) -> Option<u64> {
-    let us = |m: Microseconds| m.0.max(0) as u64;
-    match (user, system) {
-        (None, None) => None,
-        (u, s) => Some(u.map(us).unwrap_or(0) + s.map(us).unwrap_or(0)),
-    }
-}
-
-/// Assemble per-build metrics from the live-sampled cgroup snapshot (peak RAM /
-/// disk, captured before teardown) and the daemon-reported CPU time. Always
-/// reports `build_time_ms`; cgroup fields stay `None` when no cgroup was
-/// sampled (metrics disabled, idle map, or ambiguous concurrent starts).
-fn assemble_build_metrics(
-    sampled: Option<BuildMetricsRaw>,
-    cpu_usec: Option<u64>,
-    build_time_ms: u64,
-    peak_network_mbps: Option<f32>,
-) -> BuildMetrics {
-    let cpu_count = crate::metrics::host_static().cpu_count;
-    let raw = match (sampled, cpu_usec) {
-        (None, None) => None,
-        (s, cpu) => Some(BuildMetricsRaw {
-            cpu_usage_usec: cpu,
-            ..s.unwrap_or_default()
-        }),
-    };
-    if let Some(r) = raw.as_ref() {
-        let bytes = r.disk_read_bytes + r.disk_write_bytes;
-        if build_time_ms > 0 && bytes > 0 {
-            let mb_per_s = (bytes as f64 / 1_048_576.0) / (build_time_ms as f64 / 1000.0);
-            crate::metrics::throughput::DISK.observe(mb_per_s);
-        }
-    }
-    raw_to_build_metrics(raw, build_time_ms, cpu_count, peak_network_mbps)
-}
-
-/// Tracks the host's peak NAR network throughput (Mbps) over a build window.
-/// Host-level: cgroup v2 carries no per-build network accounting, so this is
-/// the closest honest signal and is exact only when the build is the sole
-/// network consumer.
-struct NetworkPeakSampler {
-    peak: std::sync::Arc<std::sync::atomic::AtomicU64>,
-    stop: std::sync::Arc<std::sync::atomic::AtomicBool>,
-    handle: tokio::task::JoinHandle<()>,
-}
-
-impl NetworkPeakSampler {
-    fn start() -> Self {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-        let peak = Arc::new(AtomicU64::new(0));
-        let stop = Arc::new(AtomicBool::new(false));
-        let (p, s) = (peak.clone(), stop.clone());
-        let handle = tokio::spawn(async move {
-            while !s.load(Ordering::Relaxed) {
-                if let Some(v) = crate::metrics::throughput::NETWORK.current() {
-                    let bits = (v as f64).to_bits();
-                    let mut prev = p.load(Ordering::Relaxed);
-                    while f64::from_bits(prev) < v as f64 {
-                        match p.compare_exchange_weak(
-                            prev,
-                            bits,
-                            Ordering::Relaxed,
-                            Ordering::Relaxed,
-                        ) {
-                            Ok(_) => break,
-                            Err(cur) => prev = cur,
-                        }
-                    }
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(250)).await;
-            }
-        });
-        Self { peak, stop, handle }
-    }
-
-    async fn finish(self) -> Option<f32> {
-        use std::sync::atomic::Ordering;
-        self.stop.store(true, Ordering::Relaxed);
-        let _ = self.handle.await;
-        match self.peak.load(Ordering::Relaxed) {
-            0 => None,
-            b => Some(f64::from_bits(b) as f32),
-        }
-    }
-}
-
-/// Samples a daemon build's cgroup `memory.peak` / `io.stat` *while it runs*,
-/// because nix destroys the cgroup as soon as the build finishes. It locates
-/// the cgroup via [`newest_build_cgroup`] (entries written after `start` was
-/// called), locks onto it, and keeps the last good reading - the high-water
-/// `memory.peak` and cumulative `io.stat` just before teardown.
-struct CgroupSampler {
-    stop: std::sync::Arc<std::sync::atomic::AtomicBool>,
-    handle: tokio::task::JoinHandle<Option<BuildMetricsRaw>>,
-}
-
-impl CgroupSampler {
-    fn start(state_dir: String, cgroup_root: String) -> Self {
-        use std::sync::Arc;
-        use std::sync::atomic::{AtomicBool, Ordering};
-        let stop = Arc::new(AtomicBool::new(false));
-        let since = std::time::SystemTime::now();
-        let s = stop.clone();
-        let handle = tokio::spawn(async move {
-            let dir = Path::new(&state_dir);
-            let mut cgroup: Option<PathBuf> = None;
-            let mut last: Option<BuildMetricsRaw> = None;
-            let mut logged = false;
-            while !s.load(Ordering::Relaxed) {
-                if cgroup.is_none() {
-                    // Trust only paths under the configured cgroup root.
-                    cgroup =
-                        newest_build_cgroup(dir, since).filter(|p| p.starts_with(&cgroup_root));
-                }
-                if let Some(dir) = cgroup.as_ref() {
-                    match read_build_cgroup(dir) {
-                        Some(cur) => {
-                            if !logged {
-                                debug!(cgroup = %dir.display(), "sampling build cgroup for metrics");
-                                logged = true;
-                            }
-                            last = Some(merge_cgroup_sample(last, cur));
-                        }
-                        None => break, // cgroup torn down; keep the last good sample
-                    }
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(CGROUP_SAMPLE_MS)).await;
-            }
-            last
-        });
-        Self { stop, handle }
-    }
-
-    async fn finish(self) -> Option<BuildMetricsRaw> {
-        use std::sync::atomic::Ordering;
-        self.stop.store(true, Ordering::Relaxed);
-        self.handle.await.ok().flatten()
-    }
-}
-
-/// Fold a fresh cgroup reading into the running snapshot: `memory.peak` is a
-/// kernel high-water mark so take the max; `io.stat` is cumulative so the latest
-/// read is the total; OOM is sticky.
-fn merge_cgroup_sample(prev: Option<BuildMetricsRaw>, cur: BuildMetricsRaw) -> BuildMetricsRaw {
-    let prev = prev.unwrap_or_default();
-    BuildMetricsRaw {
-        peak_ram_bytes: prev.peak_ram_bytes.max(cur.peak_ram_bytes),
-        cpu_usage_usec: None,
-        disk_read_bytes: cur.disk_read_bytes,
-        disk_write_bytes: cur.disk_write_bytes,
-        oom_killed: prev.oom_killed || cur.oom_killed,
-    }
-}
 
 // ── Type-state pipeline ───────────────────────────────────────────────────────
 
@@ -309,7 +76,7 @@ impl ParsedDerivation {
         let harmonia_path = StorePath::from_base_path(strip_store_prefix(&path))
             .map_err(|e| anyhow::anyhow!("invalid store path {}: {}", path, e))?;
 
-        let basic_drv = get_basic_derivation(&path, &drv)?;
+        let basic_drv = get_basic_derivation(&path, &drv).await?;
 
         Ok(Self {
             drv,
@@ -858,243 +625,9 @@ fn log_message_to_text(msg: &LogMessage) -> Option<String> {
     }
 }
 
-// ── Derivation construction ───────────────────────────────────────────────────
-
-/// Construct a harmonia [`BasicDerivation`] from a parsed drv file.
-///
-/// Output paths are taken directly from the `.drv` file:
-/// - non-empty `path` → `InputAddressed` (concrete store path)
-/// - empty `path` → `Deferred` (floating CA derivation)
-///
-/// This avoids calling `query_derivation_output_map`, which fails on some
-/// daemon versions that return full `/nix/store/...` paths where harmonia
-/// expects bare `hash-name` paths.
-///
-/// Structured attributes (`__json`) are moved from the env map to
-/// `structured_attrs` so the daemon handles them correctly.
-fn get_basic_derivation(
-    full_drv_path: &str,
-    drv: &gradient_db::Derivation,
-) -> Result<BasicDerivation> {
-    // ── Build outputs from .drv data ──────────────────────────────────────────
-    //
-    // Three shapes of `.drv` output to disambiguate:
-    //
-    // 1. Fixed-output derivation (FOD): both `hash_algo` and `hash` are
-    //    populated (e.g. a `fetchurl`). The daemon **must** see this as
-    //    `CAFixed(ContentAddress)` because that's what unlocks network
-    //    access in the build sandbox - passing it as `InputAddressed`
-    //    sandboxes it without DNS, so curl fails with
-    //    `Could not resolve host: …` and the build dies.
-    // 2. Floating CA derivation: `path` is empty AND `hash_algo` is empty.
-    //    Daemon will compute the path from the build output → `Deferred`.
-    // 3. Plain input-addressed derivation: `path` is set, no `hash_algo`.
-    //    → `InputAddressed(StorePath)`.
-    let mut outputs: BTreeMap<_, _> = BTreeMap::new();
-    for o in &drv.outputs {
-        let output_name = o
-            .name
-            .parse()
-            .with_context(|| format!("invalid output name '{}' in {}", o.name, full_drv_path))?;
-
-        let drv_output = match o.as_spec() {
-            DrvOutputSpec::FixedOutput { hash_algo, hash } => ca_fixed_output(hash_algo, hash)
-                .with_context(|| {
-                    format!(
-                        "invalid FOD spec for output '{}' in {} (hash_algo={:?} hash={:?})",
-                        o.name, full_drv_path, hash_algo, hash
-                    )
-                })?,
-            DrvOutputSpec::Deferred => DerivationOutput::Deferred,
-            DrvOutputSpec::InputAddressed { path } => {
-                let base = strip_nix_store_prefix(path);
-                let sp = StorePath::from_base_path(&base).with_context(|| {
-                    format!("invalid output path '{}' in {}", path, full_drv_path)
-                })?;
-                DerivationOutput::InputAddressed(sp)
-            }
-        };
-
-        outputs.insert(output_name, drv_output);
-    }
-
-    // ── Input paths: input_sources + output paths of input_derivations ────────
-    // The daemon needs all direct inputs present in the store before building.
-    // input_sources are plain store paths; input_derivations map drv→outputs,
-    // so we read each input .drv to resolve the concrete output paths.
-    let mut inputs: harmonia_store_path::StorePathSet = drv
-        .input_sources
-        .iter()
-        .filter_map(|p| {
-            let full = nix_store_path(p);
-            let base = strip_nix_store_prefix(&full).to_owned();
-            match StorePath::from_base_path(&base) {
-                Ok(sp) => Some(sp),
-                Err(e) => {
-                    warn!(path = %p, error = %e, "skipping input_src: not a valid store path");
-                    None
-                }
-            }
-        })
-        .collect();
-
-    for (input_drv_path, _output_names) in &drv.input_derivations {
-        let input_full = nix_store_path(input_drv_path);
-        let input_bytes = match std::fs::read(&input_full) {
-            Ok(b) => b,
-            Err(e) => {
-                warn!(drv = %input_full, error = %e, "cannot read input .drv for inputs");
-                continue;
-            }
-        };
-
-        let input_drv = match parse_drv(&input_bytes) {
-            Ok(d) => d,
-            Err(e) => {
-                warn!(drv = %input_full, error = %e, "cannot parse input .drv for inputs");
-                continue;
-            }
-        };
-
-        for o in &input_drv.outputs {
-            if o.path.is_empty() {
-                continue;
-            }
-
-            let base = strip_nix_store_prefix(&o.path);
-            match StorePath::from_base_path(&base) {
-                Ok(sp) => {
-                    inputs.insert(sp);
-                }
-
-                Err(e) => {
-                    warn!(path = %o.path, error = %e, "skipping input drv output: not a valid store path");
-                }
-            }
-        }
-    }
-
-    // ── Structured attributes ─────────────────────────────────────────────────
-    // harmonia's NixSerialize for BasicDerivation never writes `structured_attrs`
-    // to the wire - only `env` is sent. The `__json` key in env is what the Nix
-    // daemon reads for structured-attrs derivations, so leave it in place.
-
-    // Extract the name from the drv path ("hash-name.drv" → "name.drv").
-    let base = strip_nix_store_prefix(full_drv_path);
-    let drv_name = base
-        .find('-')
-        .map(|i| base[i + 1..].to_owned())
-        .unwrap_or_else(|| base.to_owned());
-
-    Ok(DerivationT {
-        name: drv_name
-            .parse()
-            .with_context(|| format!("invalid derivation name: {}", drv_name))?,
-        outputs,
-        inputs,
-        platform: Bytes::from(drv.system.clone()),
-        builder: Bytes::from(drv.builder.clone()),
-        args: drv.args.iter().map(|a| Bytes::from(a.clone())).collect(),
-        env: drv
-            .environment
-            .iter()
-            .map(|(k, v)| (Bytes::from(k.clone()), Bytes::from(v.clone())))
-            .collect(),
-        structured_attrs: None,
-    })
-}
-
-/// Build a `DerivationOutput::CAFixed(...)` from a `.drv`'s `outputHashAlgo`
-/// and `outputHash` fields. Without this the daemon would treat the FOD as
-/// an input-addressed derivation, sandbox it without network access, and
-/// every fetch (curl, git clone, …) would fail with DNS errors.
-///
-/// The `.drv` `hash_algo` field follows Nix's wire format:
-///   `"sha256"`        → flat sha256
-///   `"r:sha256"`      → recursive (NAR-hashed) sha256
-///   `"text:sha256"`   → text-hashed (rare, used by `builtins.toFile`)
-///
-/// The `hash` field is hex-encoded (base16) raw digest bytes.
-fn ca_fixed_output(hash_algo: &str, hash_hex: &str) -> Result<DerivationOutput> {
-    let (method, algo_str) = if let Some(rest) = hash_algo.strip_prefix("r:") {
-        (ContentAddressMethod::NixArchive, rest)
-    } else if let Some(rest) = hash_algo.strip_prefix("text:") {
-        (ContentAddressMethod::Text, rest)
-    } else {
-        (ContentAddressMethod::Flat, hash_algo)
-    };
-
-    let algorithm: Algorithm = algo_str
-        .parse()
-        .map_err(|e| anyhow::anyhow!("unknown hash algorithm {:?}: {}", algo_str, e))?;
-
-    let hash_bytes = hex::decode(hash_hex)
-        .with_context(|| format!("hash field {:?} is not valid hex", hash_hex))?;
-    let hash = Hash::from_slice(algorithm, &hash_bytes).with_context(|| {
-        format!(
-            "hash length {} doesn't match {:?} digest size",
-            hash_bytes.len(),
-            algorithm
-        )
-    })?;
-
-    let ca = ContentAddress::from_hash(method, hash)
-        .map_err(|e| anyhow::anyhow!("ContentAddress::from_hash failed: {}", e))?;
-    Ok(DerivationOutput::CAFixed(ca))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn ca_fixed_flat_sha256() {
-        // sha256("hello") in hex
-        let h = "2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824";
-        let out = ca_fixed_output("sha256", h).unwrap();
-        match out {
-            DerivationOutput::CAFixed(ca) => {
-                assert_eq!(ca.method(), ContentAddressMethod::Flat);
-                assert_eq!(ca.algorithm(), Algorithm::SHA256);
-            }
-            other => panic!("expected CAFixed(Flat), got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn ca_fixed_recursive_sha256() {
-        let h = "2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824";
-        let out = ca_fixed_output("r:sha256", h).unwrap();
-        match out {
-            DerivationOutput::CAFixed(ca) => {
-                assert_eq!(ca.method(), ContentAddressMethod::NixArchive);
-            }
-            other => panic!("expected CAFixed(NixArchive), got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn ca_fixed_text_sha256() {
-        let h = "2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824";
-        let out = ca_fixed_output("text:sha256", h).unwrap();
-        assert!(matches!(out, DerivationOutput::CAFixed(_)));
-    }
-
-    #[test]
-    fn ca_fixed_rejects_garbage_algo() {
-        assert!(ca_fixed_output("blake7", "deadbeef").is_err());
-    }
-
-    #[test]
-    fn ca_fixed_rejects_bad_hex() {
-        assert!(ca_fixed_output("sha256", "not-hex").is_err());
-    }
-
-    #[test]
-    fn ca_fixed_rejects_wrong_length_hash() {
-        // sha256 needs 32 bytes (64 hex chars); pass 8 bytes (16 hex chars).
-        assert!(ca_fixed_output("sha256", "deadbeefdeadbeef").is_err());
-    }
 
     fn drv_with_outputs(outputs: Vec<(&str, &str)>) -> gradient_db::Derivation {
         gradient_db::Derivation {
@@ -1257,108 +790,6 @@ mod tests {
         assert!(futures::poll!(fut.as_mut()).is_pending());
         tokio::time::advance(std::time::Duration::from_secs(6)).await;
         assert!(fut.await.is_err());
-    }
-
-    #[test]
-    fn newest_build_cgroup_ignores_entries_older_than_since() {
-        // Stale `<uid>` files (from finished builds) must not be picked: their
-        // cgroups are already gone. Only an entry written at/after the build
-        // started counts. Explicit mtimes keep this independent of filesystem
-        // timestamp granularity.
-        use std::fs::FileTimes;
-        use std::time::Duration;
-        let set_mtime = |path: &Path, t: std::time::SystemTime| {
-            std::fs::File::options()
-                .write(true)
-                .open(path)
-                .unwrap()
-                .set_times(FileTimes::new().set_modified(t))
-                .unwrap();
-        };
-        let dir = tempfile::tempdir().unwrap();
-        let since = std::time::SystemTime::now();
-
-        let stale = dir.path().join("30001");
-        std::fs::write(&stale, "/sys/fs/cgroup/nix-build-uid-30001\n").unwrap();
-        set_mtime(&stale, since - Duration::from_secs(60));
-        assert!(newest_build_cgroup(dir.path(), since).is_none());
-
-        let fresh = dir.path().join("30002");
-        std::fs::write(&fresh, "/sys/fs/cgroup/nix-build-uid-30002\n").unwrap();
-        set_mtime(&fresh, since + Duration::from_secs(60));
-        assert_eq!(
-            newest_build_cgroup(dir.path(), since),
-            Some(PathBuf::from("/sys/fs/cgroup/nix-build-uid-30002")),
-        );
-    }
-
-    #[test]
-    fn newest_build_cgroup_none_for_missing_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        let missing = dir.path().join("cgroups");
-        assert!(newest_build_cgroup(&missing, std::time::SystemTime::UNIX_EPOCH).is_none());
-    }
-
-    #[test]
-    fn daemon_cpu_usec_sums_present_fields() {
-        assert_eq!(daemon_cpu_usec(None, None), None);
-        assert_eq!(daemon_cpu_usec(Some(Microseconds(700)), None), Some(700));
-        assert_eq!(
-            daemon_cpu_usec(Some(Microseconds(700)), Some(Microseconds(300))),
-            Some(1000),
-        );
-        // Negative (unset/garbage) clamps to zero rather than underflowing.
-        assert_eq!(
-            daemon_cpu_usec(Some(Microseconds(-1)), Some(Microseconds(5))),
-            Some(5)
-        );
-    }
-
-    #[test]
-    fn raw_to_metrics_always_sets_build_time() {
-        let m = raw_to_build_metrics(None, 5_000, 4, None);
-        assert_eq!(m.build_time_ms, Some(5_000));
-        assert_eq!(m.peak_ram_mb, None);
-        assert_eq!(m.cpu_time_ms, None);
-        assert_eq!(m.avg_cpu_pct, None);
-        assert!(!m.oom_killed);
-    }
-
-    #[test]
-    fn raw_to_metrics_handles_zero_divisors() {
-        let raw = BuildMetricsRaw {
-            peak_ram_bytes: Some(2 * BYTES_PER_MB),
-            cpu_usage_usec: Some(1_000_000),
-            disk_read_bytes: 10,
-            disk_write_bytes: 20,
-            oom_killed: false,
-        };
-        // build_time_ms = 0 → avg_cpu_pct None (no divide-by-zero panic).
-        let m = raw_to_build_metrics(Some(raw), 0, 4, None);
-        assert_eq!(m.avg_cpu_pct, None);
-        // cpu_count = 0 → avg_cpu_pct None.
-        let m = raw_to_build_metrics(Some(raw), 1_000, 0, None);
-        assert_eq!(m.avg_cpu_pct, None);
-        assert_eq!(m.peak_ram_mb, Some(2));
-        assert_eq!(m.cpu_time_ms, Some(1_000));
-        assert_eq!(m.disk_read_bytes, Some(10));
-        assert_eq!(m.disk_write_bytes, Some(20));
-    }
-
-    #[test]
-    fn raw_to_metrics_computes_avg_cpu_pct() {
-        let raw = BuildMetricsRaw {
-            peak_ram_bytes: None,
-            cpu_usage_usec: Some(8_000_000), // 8000 ms of CPU
-            disk_read_bytes: 0,
-            disk_write_bytes: 0,
-            oom_killed: false,
-        };
-        // 8000 cpu-ms over 4000 ms wall on 4 cores = 8000/(4000*4)*100 = 50%.
-        let m = raw_to_build_metrics(Some(raw), 4_000, 4, Some(125.0));
-        assert_eq!(m.cpu_time_ms, Some(8_000));
-        assert_eq!(m.avg_cpu_pct, Some(50.0));
-        assert_eq!(m.peak_network_mbps, Some(125.0));
     }
 
     #[tokio::test]
